requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-docx==1.1.0
markdown==3.5.1
requests==2.31.0
pillow==10.1.0
